@@ -1,17 +1,19 @@
 [package]
+edition = "2024"
 name = "spawn_server"
 version = "0.1.0"
-edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-
 [dependencies]
-actix-web = "*"
-serde = "*"
 actix-rt = "*"
+actix-web = "*"
+reqwest = {version = "*", features = ["native-tls", "json", "blocking"]}
+serde = "*"
 serde_derive = "*"
 serde_json = "*"
 shells = "*"
-reqwest = { version = "*", features = ["native-tls", "json", "blocking"] }
-tokio = { version = "*", features = ["full"] }
+tokio = {version = "*", features = ["full"]}
+tracing = "*"
+tracing-actix-web = "*"
+tracing-subscriber = {version = "*", features = ["env-filter", "fmt"]}
