@@ -1,5 +1,9 @@
 use serde_derive::{Deserialize, Serialize};
+use std::net::{IpAddr, SocketAddr};
+use tracing::{error, info, warn};
 
+const DEFAULT_SPAWN_SERVER_PORT: u16 = 8099;
+const DEFAULT_SPAWN_SERVER_HOST: &str = "127.0.0.1";
 #[derive(Debug, Serialize, Deserialize, Default)]
 pub struct Command {
     pub command: String,
@@ -12,13 +16,40 @@ pub struct CommandResponse {
     pub stderr: String,
 }
 
-fn get_spawner_url() -> String {
-    "http://127.0.0.1:8099/command".to_string()
+pub fn get_spawn_server_addr() -> SocketAddr {
+    // Get host string
+    let host = std::env::var("SPAWN_SERVER_HOST")
+        .unwrap_or_else(|_| DEFAULT_SPAWN_SERVER_HOST.to_string());
+
+    // Parse host as IP
+    let ip: IpAddr = host
+        .parse()
+        .unwrap_or_else(|_| {
+            eprintln!(
+                "Invalid SPAWN_SERVER_HOST value: {host}. Must be a valid IP address (e.g., 127.0.0.1 or 0.0.0.0)"
+            );
+            std::process::exit(1);
+        });
+
+    // Parse port
+    let port = std::env::var("SPAWN_SERVER_PORT")
+        .unwrap_or_else(|_| DEFAULT_SPAWN_SERVER_PORT.to_string())
+        .parse::<u16>()
+        .unwrap_or_else(|_| {
+            eprintln!("Invalid SPAWN_SERVER_PORT value. Must be a number between 1 and 65535.");
+            std::process::exit(1);
+        });
+
+    SocketAddr::new(ip, port)
 }
 
-pub fn sync_remote_execute(cmd: &str) -> (i32, String, String) {
+fn get_spawner_command_url() -> String {
+    let server_addr = get_spawn_server_addr();
+    format!("{server_addr}/command")
+}
+pub fn sync_remote_execute<T: AsRef<str>>(cmd: T) -> (i32, String, String) {
     let cmd = Command {
-        command: cmd.to_string(),
+        command: cmd.as_ref().to_string(),
     };
 
     use std::time::Duration;
@@ -27,113 +58,112 @@ pub fn sync_remote_execute(cmd: &str) -> (i32, String, String) {
     let client_result = reqwest::blocking::Client::builder()
         .timeout(very_long_timeout)
         .build();
+
     let client = match client_result {
         Ok(c) => c,
         Err(e) => {
-            eprintln!("sync_remote_execute: Failed to build client: {}", e);
-            return (
-                -3,
-                "".to_string(),
-                format!("sync_remote_execute API: Client Build Error (ERROR 910-21087-27552). Error: {e}"),
-            );
+            error!(error = %e, "Failed to build reqwest client");
+            return (-3, "".to_string(), format!("Client Build Error: {e}"));
         }
     };
-
-    match client.post(get_spawner_url()).json(&cmd).send() {
+    let spawn_server_command_url = get_spawner_command_url();
+    match client.post(spawn_server_command_url).json(&cmd).send() {
         Ok(resp) => {
             if resp.status().is_success() {
-                let result: CommandResponse = resp.json().unwrap();
-                eprintln!("sync_remote_execute API: Success - received code {}", result.code);
-                (result.code, result.stdout, result.stderr)
+                match resp.json::<CommandResponse>() {
+                    Ok(result) => {
+                        info!(code = result.code, "sync command executed successfully");
+                        (result.code, result.stdout, result.stderr)
+                    }
+                    Err(e) => {
+                        error!(error = %e, "Failed to parse success response JSON");
+                        (-4, "".to_string(), format!("JSON parse error: {e}"))
+                    }
+                }
             } else {
-                let result: CommandResponse = resp.json().unwrap();
-                eprintln!("sync_remote_execute API: return status indicates no success (ERROR 8192-173-10620)");
-                (
-                    -2,
-                    format!("{}", result.stdout),
-                    format!("sync_remote_execute API: No Success Error (ERROR 8192-173-10620): {}", result.stderr),
-                )
+                warn!(status = %resp.status(), "sync command returned non-success status");
+                match resp.json::<CommandResponse>() {
+                    Ok(result) => (
+                        -2,
+                        result.stdout,
+                        format!("No Success Error: {}", result.stderr),
+                    ),
+                    Err(e) => {
+                        error!(error = %e, "Failed to parse error response JSON");
+                        (-5, "".to_string(), format!("JSON parse error: {e}"))
+                    }
+                }
             }
         }
         Err(e) => {
-            eprintln!(
-                "sync_remote_execute API: response cannot be parsed! {} (ERROR 67132-2323-78223)",
-                e
-            );
-            (
-                -1,
-                "".to_string(),
-                format!("sync_remote_execute API: RPC Error (ERROR 67132-2323-78224): {}",  e),
-            )
+            error!(error = %e, "sync command request failed");
+            (-1, "".to_string(), format!("RPC Error: {e}"))
         }
     }
 }
 
-//  async_version
-
-pub async fn remote_execute(cmd: &str) -> (i32, String, String) {
+pub async fn async_remote_execute<T: AsRef<str>>(cmd: T) -> (i32, String, String) {
     let cmd = Command {
-        command: cmd.to_string(),
+        command: cmd.as_ref().to_owned(),
     };
 
     use std::time::Duration;
     let very_long_timeout = Duration::new(60 * 60 * 24, 0);
 
-    let client = reqwest::Client::builder()
+    let client = match reqwest::Client::builder()
         .timeout(very_long_timeout)
         .build()
-        .unwrap(); // todo: remove unwrap
+    {
+        Ok(c) => c,
+        Err(e) => {
+            error!(error = %e, "Failed to build async reqwest client");
+            return (-3, "".to_string(), format!("Client Build Error: {e}"));
+        }
+    };
+    let spawn_server_command_url = get_spawner_command_url();
 
-    match client.post(get_spawner_url()).json(&cmd).send().await {
+    match client
+        .post(spawn_server_command_url)
+        .json(&cmd)
+        .send()
+        .await
+    {
         Ok(resp) => {
             if resp.status().is_success() {
-                let result: CommandResponse = resp.json().await.unwrap();
-                (result.code, result.stdout, result.stderr)
+                match resp.json::<CommandResponse>().await {
+                    Ok(result) => {
+                        info!(code = result.code, "async command executed successfully");
+                        (result.code, result.stdout, result.stderr)
+                    }
+                    Err(e) => {
+                        error!(error = %e, "Failed to parse async success response JSON");
+                        (-4, "".to_string(), format!("JSON parse error: {e}"))
+                    }
+                }
             } else {
-                eprintln!("async_remote_execute API: return status indicates no success (ERROR 8192-173-10620)");
-                (
-                    -2,
-                    "".to_string(),
-                    "No Success Error (ERROR 8192-173-10620)".to_string(),
-                )
+                warn!(status = %resp.status(), "async command returned non-success status");
+                (-2, "".to_string(), "No Success Error".to_string())
             }
         }
         Err(e) => {
-            eprintln!(
-                "async_remote_execute API response cannot be parsed! {} (ERROR 67132-2323-78123)",
-                e
-            );
-            (
-                -1,
-                "".to_string(),
-                "RPC Error  (ERROR 67132-2323-78124)".to_string(),
-            )
+            error!(error = %e, "async command request failed");
+            (-1, "".to_string(), format!("RPC Error: {e}"))
         }
     }
 }
 
 /// Macro to execute the given command on the spawn server using synchronous communcation
-
 #[macro_export]
 macro_rules! srpc {
     ( $( $cmd:tt )* ) => {{
-        $crate::sync_remote_execute(&format!($( $cmd )*))
+        $crate::sync_remote_execute(format!($( $cmd )*))
     }};
 }
 
 /// Macro to execute the given command on the spawn server using asynchronous communcation
-
 #[macro_export]
 macro_rules! arpc {
     ( $( $cmd:tt )* ) => {{
-        $crate::remote_execute(&format!($( $cmd )*))
-    }};
-}
-
-
-#[macro_export]
-macro_rules! sh {
-    ( $( $cmd:tt )* ) => {{
-        $crate::sync_remote_execute(&format!($( $cmd )*))
+        $crate::async_remote_execute(format!($( $cmd )*))
     }};
 }
