@@ -1,17 +1,25 @@
-use actix_web::get;
-use actix_web::{post, web::Json, App, HttpResponse, HttpServer, Responder};
+use actix_web::{App, HttpResponse, HttpServer, Responder, get, post, web::Json};
 use shells::sh;
-use spawn_server::{Command, CommandResponse};
-use tokio::task;
+use spawn_server::{Command, CommandResponse, get_spawn_server_addr};
+use tracing::{error, info, level_filters::LevelFilter, warn};
+use tracing_actix_web::TracingLogger;
+use tracing_subscriber::EnvFilter;
 
 #[post("/command")]
 async fn info(command: Json<Command>) -> impl Responder {
     let cmd = command.command.clone();
+
+    if command.command.trim().is_empty() {
+        return HttpResponse::BadRequest().body("command must not be empty");
+    }
+
     let response = if let Ok((code, stdout, stderr)) =
-        task::spawn_blocking(move || sh!("{}", command.command)).await
+        tokio::task::spawn_blocking(move || sh!("{}", command.command)).await
     {
         if code != 0 {
-            eprintln!("{cmd} failed: stdout='{stdout}', stderr='{stderr}'");
+            warn!(%cmd, %stdout, %stderr, "command failed");
+        } else {
+            info!(%cmd, "command executed successfully");
         }
         CommandResponse {
             code,
@@ -19,38 +27,41 @@ async fn info(command: Json<Command>) -> impl Responder {
             stderr,
         }
     } else {
-        eprintln!("{cmd}: failed to spawn command.");
+        error!(%cmd, "failed to spawn command");
         CommandResponse {
             code: 100,
-            stdout: format!("spawn_server: command '{cmd}' failed (ERROR 128912-12128-18492)"),
-            stderr: "(ERROR 128912-12128-18493)".to_string(),
+            stdout: format!("spawn_server: command '{cmd}' failed"),
+            stderr: "spawn error".to_string(),
         }
     };
+
     HttpResponse::Ok().json(response)
 }
 
 #[get("/")]
 async fn index() -> impl Responder {
-    HttpResponse::Ok().body(r#"{"server": "spawn_server", "version": "0.1.0"}"#)
+    HttpResponse::Ok().json(r#"{"server": "spawn_server", "version": "0.1.0"}"#)
 }
 
-fn main() {
-    let _ = actix_web::rt::System::with_tokio_rt(|| {
-        tokio::runtime::Builder::new_multi_thread()
-            .enable_all()
-            .worker_threads(16)
-            .thread_name("main-tokio")
-            .build()
-            .unwrap()
+#[actix_web::main]
+async fn main() -> std::io::Result<()> {
+    // initialize tracing
+    tracing_subscriber::fmt()
+        .with_env_filter(EnvFilter::from_default_env().add_directive(LevelFilter::INFO.into())) // RUST_LOG controls level, info as default
+        .with_target(false)
+        .init();
+
+    let addr = get_spawn_server_addr();
+
+    tracing::info!("Starting spawn_server on {addr}");
+
+    HttpServer::new(|| {
+        App::new()
+            .wrap(TracingLogger::default()) // <-- per-request spans
+            .service(index)
+            .service(info)
     })
-    .block_on(async_main());
-}
-
-async fn async_main() -> std::io::Result<()> {
-    HttpServer::new(|| App::new().service(index).service(info))
-        .workers(16)
-        .bind("127.0.0.1:8099")?
-        .run()
-        .await?;
-    Ok(())
+    .bind(addr)?
+    .run()
+    .await
 }
